バリデーションとアノテーションの設定。バリデーションメッセージの表示

+++ src/main/java/com/example/demo/entity/User.java
@@ -22,7 +22,6 @@ import jakarta.persistence.OneToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
-
 import lombok.Getter;
 import lombok.Setter;
 
@@ -39,7 +38,7 @@ public class User {
     private String password;
 
     @Column(name = "employee_no")
-    private Integer employeeNo;
+    private Long employeeNo;
 
     @Column(name = "joining_date", nullable = false)
     private LocalDate joiningDate;
