+++ src/main/java/com/example/demo/service/UserService.java
@@ -1,7 +1,5 @@
 package com.example.demo.service;
 
-import java.time.LocalDate;
-import java.util.List;
 import java.util.Optional;
 
 import com.example.demo.entity.User;
@@ -16,5 +14,5 @@ public interface UserService {
 
     User findByEmail(String email);
 
-    User findByEmployeeNo(Integer employeeNo);
+    User findByEmployeeNo(Long employeeNo);
 }
